requirements.txt
langchain>=0.2.9
langchain-openai>=0.1.8
langchain-community>=0.2.7
openai>=1.40.0
chromadb>=0.4.24
python-dotenv>=1.0.1
rich>=13.7.1
